# docs/swagger.yaml
definitions:
  controllers.CreateArticleRequest:
    properties:
      content:
        example: 文章内容
        type: string
      tags:
        example:
        - '["标签1"'
        - '"标签2"]'
        items:
          type: string
        type: array
      title:
        example: 文章标题
        type: string
    required:
    - content
    - title
    type: object
  controllers.CreateCommentRequest:
    properties:
      content:
        example: 评论内容
        type: string
      parent_id:
        example: 1
        type: integer
    required:
    - content
    type: object
  controllers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.RegisterRequest:
    properties:
      email:
        type: string
      nickname:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.UpdateArticleRequest:
    properties:
      content:
        example: 更新后的内容
        type: string
      tags:
        example:
        - '["标签1"'
        - '"标签2"]'
        items:
          type: string
        type: array
      title:
        example: 更新后的标题
        type: string
    type: object
  controllers.WechatLoginRequest:
    properties:
      avatarUrl:
        type: string
      city:
        type: string
      code:
        type: string
      country:
        type: string
      gender:
        type: integer
      language:
        type: string
      nickname:
        type: string
      province:
        type: string
    required:
    - code
    type: object
  models.Article:
    properties:
      author:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 作者信息
      author_id:
        description: 作者ID
        type: integer
      comment_count:
        description: 评论数
        type: integer
      comments:
        description: 评论列表
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      content:
        description: 文章内容
        type: string
      created_at:
        type: string
      id:
        type: integer
      like_count:
        description: 点赞数
        type: integer
      likes:
        description: 点赞列表
        items:
          $ref: '#/definitions/models.Like'
        type: array
      tags:
        description: 文章标签
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      title:
        description: 文章标题
        type: string
      updated_at:
        type: string
    type: object
  models.Comment:
    properties:
      article:
        allOf:
        - $ref: '#/definitions/models.Article'
        description: 评论文章
      article_id:
        description: 评论文章ID
        type: integer
      children:
        description: 子评论列表
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      content:
        description: 评论内容
        type: string
      created_at:
        type: string
      id:
        type: integer
      parent:
        allOf:
        - $ref: '#/definitions/models.Comment'
        description: 父评论
      parent_id:
        description: 父评论ID，用于实现树状结构
        type: integer
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 评论用户
      user_id:
        description: 评论用户ID
        type: integer
    type: object
  models.Like:
    properties:
      article:
        allOf:
        - $ref: '#/definitions/models.Article'
        description: 被点赞文章
      article_id:
        description: 被点赞文章ID
        type: integer
      created_at:
        type: string
      id:
        type: integer
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 点赞用户
      user_id:
        description: 点赞用户ID
        type: integer
    type: object
  models.Tag:
    properties:
      articles:
        description: 关联的文章
        items:
          $ref: '#/definitions/models.Article'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        description: 标签名称
        type: string
      updated_at:
        type: string
    type: object
  models.User:
    description: 用户信息
    properties:
      avatar:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        description: 创建时间
        type: string
      deleted_at:
        description: 删除时间
        type: string
      email:
        type: string
      followers_count:
        description: 论坛社交相关字段
        type: integer
      following_count:
        type: integer
      gender:
        type: integer
      id:
        description: 用户ID
        type: integer
      language:
        type: string
      last_login_at:
        description: 额外的时间字段
        type: string
      nickname:
        type: string
      open_id:
        description: 微信小程序相关字段
        type: string
      password:
        type: string
      phone:
        type: string
      posts_count:
        type: integer
      province:
        type: string
      session_key:
        type: string
      union_id:
        type: string
      updated_at:
        description: 更新时间
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/articles:
    get:
      description: 获取文章列表，支持分页
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Article'
            type: array
      summary: 获取文章列表
      tags:
      - 文章管理
    post:
      consumes:
      - application/json
      description: 创建一篇新文章
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章信息
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Article'
      summary: 创建文章
      tags:
      - 文章管理
  /api/articles/{id}:
    delete:
      description: 删除指定的文章
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 删除文章
      tags:
      - 文章管理
    get:
      description: 获取指定文章的详细信息
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Article'
      summary: 获取文章详情
      tags:
      - 文章管理
    put:
      consumes:
      - application/json
      description: 更新指定文章的信息
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文章更新信息
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Article'
      summary: 更新文章
      tags:
      - 文章管理
  /api/articles/{id}/comments:
    post:
      consumes:
      - application/json
      description: 为指定文章添加评论
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论信息
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
      summary: 评论文章
      tags:
      - 文章管理
  /api/articles/{id}/like:
    post:
      description: 为指定文章点赞或取消点赞
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 点赞文章
      tags:
      - 文章管理
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口
      parameters:
      - description: 登录参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 返回用户信息和token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 认证失败
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 用户注册接口
      parameters:
      - description: 注册参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 返回用户信息和token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      summary: 用户注册
      tags:
      - 认证
  /auth/wechat/login:
    post:
      consumes:
      - application/json
      description: 通过微信小程序登录接口，支持新用户注册和老用户登录
      parameters:
      - description: 微信登录参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.WechatLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 微信小程序登录
      tags:
      - 认证
swagger: "2.0"